Refactor: Restructure controllers for clarity and responsibility separation
