+++ pentasharp/Controllers/BusScheduleCompanyController.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace WebApplication1.Controllers
+{
+	public class BusScheduleCompanyController : Controller
+	{
+		public IActionResult AddBusCompany()
+		{
+			return View();
+		}
+	}
+}
