+++ pentasharp/Controllers/BusScheduleController.cs
@@ -15,16 +15,6 @@ namespace WebApplication1.Controllers
             return View();
         }
 
-        public IActionResult AddBusCompany()
-        {
-            return View(); 
-        }
-
-        public IActionResult ConfirmPassengers()
-        {
-            return View();
-        }
-
         public IActionResult ManageBusSchedules()
         {
             return View();
