+++ pentasharp/Controllers/BusSchedulePassenger.cs
@@ -2,9 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace WebApplication1.Controllers
 {
-    public class TaxiReservationController : Controller
+    public class BusSchedulePassengerController : Controller
     {
-        public IActionResult TaxiReservations()
+        public IActionResult ConfirmPassengers()
         {
             return View();
         }
