+++ pentasharp/Controllers/TaxiCompanyBookingController.cs
@@ -2,9 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace WebApplication1.Controllers
 {
-    public class TaxiReservationController : Controller
+    public class TaxiCompanyBookingController : Controller
     {
-        public IActionResult TaxiReservations()
+        public IActionResult AddTaxiCompany()
         {
             return View();
         }
