+++ pentasharp/Views/BusReservation/BusReservationsManagement.cshtml
@@ -14,7 +14,7 @@
                     </div>
                     <h5 class="card-title text-primary fw-bold">Confirm Reservations</h5>
                     <p class="card-text text-muted">Review and confirm pending reservations for bus trips.</p>
-                    <a href="@Url.Action("ConfirmPassengers", "BusReservation")" class="btn btn-outline-primary mt-3">Confirm Reservations</a>
+                    <a href="@Url.Action("ConfirmPassengers", "BusReservationPassenger")" class="btn btn-outline-primary mt-3">Confirm Reservations</a>
                 </div>
             </div>
         </div>
@@ -27,7 +27,7 @@
                     </div>
                     <h5 class="card-title text-primary fw-bold">View Passengers by Route</h5>
                     <p class="card-text text-muted">View all confirmed passengers sorted by bus route and schedule.</p>
-                    <a href="@Url.Action("ViewPassangers", "BusReservation")" class="btn btn-outline-primary mt-3">View Passengers</a>
+                    <a href="@Url.Action("ViewPassengers", "BusReservationPassenger")" class="btn btn-outline-primary mt-3">View Passengers</a>
                 </div>
             </div>
         </div>
