+++ pentasharp/Views/BusSchedule/BusScheduleManagement.cshtml
@@ -14,7 +14,7 @@
                     </div>
                     <h5 class="card-title text-primary fw-bold">Confirm Passengers</h5>
                     <p class="card-text text-muted">Review and confirm passengers for scheduled trips.</p>
-                    <a href="@Url.Action("ConfirmPassengers", "BusSchedule")" class="btn btn-outline-primary">Confirm Passengers</a>
+                    <a href="@Url.Action("ConfirmPassengers", "BusSchedulePassenger")" class="btn btn-outline-primary">Confirm Passengers</a>
                 </div>
             </div>
         </div>
