+++ pentasharp/Views/Dashboard/Dashboard.cshtml
@@ -39,7 +39,7 @@
                     </div>
                     <h6 class="card-title">Taxi Bookings</h6>
                     <p class="stat-number">120</p>
-                    <a href="@Url.Action("TaxiBookingsManagement", "TaxiBooking")" class="dashboard-link mt-auto">Manage</a>
+                    <a href="@Url.Action("TaxiBookingsManagement", "TaxiBookingManagement")" class="dashboard-link mt-auto">Manage</a>
                 </div>
             </div>
         </div>
@@ -52,7 +52,7 @@
                     </div>
                     <h6 class="card-title">Taxi Reservations</h6>
                     <p class="stat-number">80</p>
-                    <a href="@Url.Action("TaxiReservationsManagement", "TaxiReservation")" class="dashboard-link mt-auto">Manage</a>
+                    <a href="@Url.Action("TaxiReservationsManagement", "TaxiReservationManagement")" class="dashboard-link mt-auto">Manage</a>
                 </div>
             </div>
         </div>
@@ -79,7 +79,7 @@
                         <i class="fas fa-building fa-2x"></i>
                     </div>
                     <h6 class="card-title">Add New Company and Bus</h6>
-                    <a href="@Url.Action("AddBusCompany", "BusSchedule")" class="dashboard-link mt-auto">Add</a>
+                    <a href="@Url.Action("AddBusCompany", "BusScheduleCompany")" class="dashboard-link mt-auto">Add</a>
                 </div>
             </div>
         </div>
@@ -98,7 +98,7 @@
                         <i class="fas fa-building fa-2x"></i>
                     </div>
                     <h6 class="card-title">Add New Company and Taxi</h6>
-                    <a href="@Url.Action("AddTaxiCompany", "TaxiBooking")" class="dashboard-link mt-auto">Add</a>
+                    <a href="@Url.Action("AddTaxiCompany", "TaxiCompanyBooking")" class="dashboard-link mt-auto">Add</a>
                 </div>
             </div>
         </div>
