--- pentasharp/bin/Debug/net8.0/appsettings.json
@@ -1,15 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    //   "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=PentasharpDB;Trusted_Connection=True;MultipleActiveResultSets=true"
-    //  "DefaultConnection": "Server=LAPTOP-BHPP930E;Database=PentasharpDB;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True",
-    "DefaultConnection": "Server=EDONS_LAPTOP;Database=PentasharpDB;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True"
-  }
-}
-
